verse formatting + removed all codfe from elisha
