+++ frontend/app/lib/pages/bible_reader/flowing_chapter_text.dart
@@ -6,7 +6,7 @@
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:app/pages/bible_reader/bible_repo_elisha.dart';
+import 'package:app/models/bible_reader/verse_ref.dart';
 import 'package:app/models/bible_reader/highlight.dart'; // for HighlightColor
 
 /// The following elements are stored in this class:
